+++ x/gov/keeper/grpc_query.go
@@ -133,14 +133,18 @@ func (q Keeper) Votes(c context.Context, req *types.QueryVotesRequest) (*types.Q
 		return nil, status.Error(codes.InvalidArgument, "proposal id can not be 0")
 	}
 
-	var votes types.Votes
+	var votes govtypes.Votes
 	ctx := sdk.UnwrapSDKContext(c)
 
 	store := ctx.KVStore(q.storeKey)
 	votesStore := prefix.NewStore(store, govtypes.VotesKey(req.ProposalId))
 
 	pageRes, err := query.Paginate(votesStore, req.Pagination, func(key []byte, value []byte) error {
+<<<<<<< HEAD
 		var vote types.Vote
+=======
+		var vote govtypes.Vote
+>>>>>>> 6f4b45bce5f277e193c4116dbea18212f40e242a
 		if err := q.cdc.Unmarshal(value, &vote); err != nil {
 			return err
 		}
@@ -222,14 +226,18 @@ func (q Keeper) Deposits(c context.Context, req *types.QueryDepositsRequest) (*t
 		return nil, status.Error(codes.InvalidArgument, "proposal id can not be 0")
 	}
 
-	var deposits types.Deposits
+	var deposits govtypes.Deposits
 	ctx := sdk.UnwrapSDKContext(c)
 
 	store := ctx.KVStore(q.storeKey)
 	depositStore := prefix.NewStore(store, govtypes.DepositsKey(req.ProposalId))
 
 	pageRes, err := query.Paginate(depositStore, req.Pagination, func(key []byte, value []byte) error {
+<<<<<<< HEAD
 		var deposit types.Deposit
+=======
+		var deposit govtypes.Deposit
+>>>>>>> 6f4b45bce5f277e193c4116dbea18212f40e242a
 		if err := q.cdc.Unmarshal(value, &deposit); err != nil {
 			return err
 		}