+++ x/gov/keeper/keeper.go
@@ -107,13 +107,17 @@ func (k Keeper) IterateInactiveProposalsQueue(ctx sdk.Context, endTime time.Time
 }
 
 // IterateAllDeposits iterates over the all the stored deposits and performs a callback function.
-func (k Keeper) IterateAllDeposits(ctx sdk.Context, cb func(deposit types.Deposit) (stop bool)) {
+func (k Keeper) IterateAllDeposits(ctx sdk.Context, cb func(deposit govtypes.Deposit) (stop bool)) {
 	store := ctx.KVStore(k.storeKey)
 	iterator := sdk.KVStorePrefixIterator(store, govtypes.DepositsKeyPrefix)
 
 	defer iterator.Close()
 	for ; iterator.Valid(); iterator.Next() {
+<<<<<<< HEAD
 		var deposit types.Deposit
+=======
+		var deposit govtypes.Deposit
+>>>>>>> 6f4b45bce5f277e193c4116dbea18212f40e242a
 		k.cdc.MustUnmarshal(iterator.Value(), &deposit)
 
 		if cb(deposit) {