+++ x/gov/keeper/tally.go
@@ -28,7 +28,7 @@ func Tally(ctx sdk.Context, k Keeper, proposal types.Proposal) (pass bool, veto 
 
 	fetchBondedValidators(ctx, k, currValidators)
 
-	k.IterateVotes(ctx, proposal.ProposalId, func(vote types.Vote) bool {
+	k.IterateVotes(ctx, proposal.ProposalId, func(vote govtypes.Vote) bool {
 		voter, err := sdk.AccAddressFromBech32(vote.Voter)
 		if err != nil {
 			panic(err)
@@ -130,7 +130,11 @@ func fetchBondedValidators(ctx sdk.Context, k Keeper, validators map[string]*val
 	})
 }
 
+<<<<<<< HEAD
 func delegatorVoting(ctx sdk.Context, k Keeper, vote types.Vote, validators map[string]*validatorGovInfo, results map[govtypes.VoteOption]sdk.Dec, totalVotingPower *sdk.Dec) {
+=======
+func delegatorVoting(ctx sdk.Context, k Keeper, vote govtypes.Vote, validators map[string]*validatorGovInfo, results map[govtypes.VoteOption]sdk.Dec, totalVotingPower *sdk.Dec) {
+>>>>>>> 6f4b45bce5f277e193c4116dbea18212f40e242a
 	voter, err := sdk.AccAddressFromBech32(vote.Voter)
 	if err != nil {
 		panic(err)
@@ -268,10 +272,15 @@ func SecurityTally(ctx sdk.Context, k Keeper, proposal types.Proposal) (bool, bo
 
 	currVotes := k.GetVotes(ctx, proposal.ProposalId)
 	for _, vote := range currVotes {
+<<<<<<< HEAD
 		if vote.Option == govtypes.OptionEmpty {
+=======
+		if len(vote.Options) != 1 {
+>>>>>>> 6f4b45bce5f277e193c4116dbea18212f40e242a
 			continue
 		}
-		results[vote.Option] = results[vote.Option].Add(sdk.NewDec(1))
+
+		results[vote.Options[0].Option] = results[vote.Options[0].Option].Add(sdk.NewDec(1))
 		totalHeadCounts = totalHeadCounts.Add(sdk.NewDec(1))
 	}
 	tallyParams := k.GetTallyParams(ctx)