+++ x/gov/keeper/vote.go
@@ -6,7 +6,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	govTypes "github.com/cosmos/cosmos-sdk/x/gov/types"
+	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 	"github.com/tendermint/tendermint/crypto/tmhash"
 
 	"github.com/certikfoundation/shentu/v2/x/gov/types"
@@ -14,17 +14,22 @@ import (
 )
 
 // AddVote Adds a vote on a specific proposal.
+<<<<<<< HEAD
 func (k Keeper) AddVote(ctx sdk.Context, proposalID uint64, voterAddr sdk.AccAddress, options govTypes.WeightedVoteOptions) error {
+=======
+func (k Keeper) AddVote(ctx sdk.Context, proposalID uint64, voterAddr sdk.AccAddress, options govtypes.WeightedVoteOptions) error {
+>>>>>>> 6f4b45bce5f277e193c4116dbea18212f40e242a
 	proposal, ok := k.GetProposal(ctx, proposalID)
 	if !ok {
-		return sdkerrors.Wrapf(govTypes.ErrUnknownProposal, "%v", proposalID)
+		return sdkerrors.Wrapf(govtypes.ErrUnknownProposal, "%v", proposalID)
 	}
 	if proposal.Status != types.StatusCertifierVotingPeriod &&
 		proposal.Status != types.StatusValidatorVotingPeriod {
-		return sdkerrors.Wrapf(govTypes.ErrInactiveProposal, "%v", proposalID)
+		return sdkerrors.Wrapf(govtypes.ErrInactiveProposal, "%v", proposalID)
 	}
 
 	for _, option := range options {
+<<<<<<< HEAD
 		if !govTypes.ValidWeightedVoteOption(option) {
 			return sdkerrors.Wrapf(govTypes.ErrInvalidVote, "%s", option)
 		}
@@ -33,30 +38,50 @@ func (k Keeper) AddVote(ctx sdk.Context, proposalID uint64, voterAddr sdk.AccAdd
 			if !(option.Option == govTypes.OptionYes ||
 				option.Option == govTypes.OptionNo) {
 				return sdkerrors.Wrapf(govTypes.ErrInvalidVote,
+=======
+		if !govtypes.ValidWeightedVoteOption(option) {
+			return sdkerrors.Wrapf(govtypes.ErrInvalidVote, "%s", option)
+		}
+
+		if proposal.Status == types.StatusCertifierVotingPeriod {
+			if !(option.Option == govtypes.OptionYes ||
+				option.Option == govtypes.OptionNo) {
+				return sdkerrors.Wrapf(govtypes.ErrInvalidVote,
+>>>>>>> 6f4b45bce5f277e193c4116dbea18212f40e242a
 					"'%s' is not valid option in certifier voting; must be 'yes' or 'no'", option.Option)
 			}
 		}
 	}
 
 	if proposal.Status == types.StatusCertifierVotingPeriod && !k.IsCertifier(ctx, voterAddr) {
-		return sdkerrors.Wrapf(govTypes.ErrInvalidVote, "'%s' is not a certifier.", voterAddr)
+		return sdkerrors.Wrapf(govtypes.ErrInvalidVote, "'%s' is not a certifier.", voterAddr)
 	}
 
 	if proposal.GetContent().ProposalType() == shieldtypes.ProposalTypeShieldClaim &&
 		proposal.Status == types.StatusValidatorVotingPeriod &&
 		!k.IsCertifiedIdentity(ctx, voterAddr) {
-		return sdkerrors.Wrapf(govTypes.ErrInvalidVote, "'%s' is not a certified identity", voterAddr)
+		return sdkerrors.Wrapf(govtypes.ErrInvalidVote, "'%s' is not a certified identity", voterAddr)
 	}
 
 	txhash := hex.EncodeToString(tmhash.Sum(ctx.TxBytes()))
+<<<<<<< HEAD
 	vote := types.NewVote(proposalID, voterAddr, options, txhash)
+=======
+	vote := govtypes.NewVote(proposalID, voterAddr, options)
+>>>>>>> 6f4b45bce5f277e193c4116dbea18212f40e242a
 	k.setVote(ctx, vote)
 
 	ctx.EventManager().EmitEvent(
 		sdk.NewEvent(
+<<<<<<< HEAD
 			govTypes.EventTypeProposalVote,
 			sdk.NewAttribute(govTypes.AttributeKeyOption, options.String()),
 			sdk.NewAttribute(govTypes.AttributeKeyProposalID, fmt.Sprintf("%d", proposalID)),
+=======
+			govtypes.EventTypeProposalVote,
+			sdk.NewAttribute(govtypes.AttributeKeyOption, options.String()),
+			sdk.NewAttribute(govtypes.AttributeKeyProposalID, fmt.Sprintf("%d", proposalID)),
+>>>>>>> 6f4b45bce5f277e193c4116dbea18212f40e242a
 			sdk.NewAttribute(types.AttributeKeyVoter, voterAddr.String()),
 			sdk.NewAttribute(types.AttributeTxHash, txhash),
 		),
@@ -66,8 +91,8 @@ func (k Keeper) AddVote(ctx sdk.Context, proposalID uint64, voterAddr sdk.AccAdd
 }
 
 // GetAllVotes returns all the votes from the store.
-func (k Keeper) GetAllVotes(ctx sdk.Context) (votes types.Votes) {
-	k.IterateAllVotes(ctx, func(vote types.Vote) bool {
+func (k Keeper) GetAllVotes(ctx sdk.Context) (votes govtypes.Votes) {
+	k.IterateAllVotes(ctx, func(vote govtypes.Vote) bool {
 		votes = append(votes, vote)
 		return false
 	})
@@ -75,8 +100,8 @@ func (k Keeper) GetAllVotes(ctx sdk.Context) (votes types.Votes) {
 }
 
 // GetVotes returns all votes on a given proposal.
-func (k Keeper) GetVotes(ctx sdk.Context, proposalID uint64) (votes types.Votes) {
-	k.IterateVotes(ctx, proposalID, func(vote types.Vote) bool {
+func (k Keeper) GetVotes(ctx sdk.Context, proposalID uint64) (votes govtypes.Votes) {
+	k.IterateVotes(ctx, proposalID, func(vote govtypes.Vote) bool {
 		votes = append(votes, vote)
 		return false
 	})
@@ -84,8 +109,8 @@ func (k Keeper) GetVotes(ctx sdk.Context, proposalID uint64) (votes types.Votes)
 }
 
 // GetVotesPaginated performs paginated query of votes on a given proposal.
-func (k Keeper) GetVotesPaginated(ctx sdk.Context, proposalID uint64, page, limit uint) (votes types.Votes) {
-	k.IterateVotesPaginated(ctx, proposalID, page, limit, func(vote types.Vote) bool {
+func (k Keeper) GetVotesPaginated(ctx sdk.Context, proposalID uint64, page, limit uint) (votes govtypes.Votes) {
+	k.IterateVotesPaginated(ctx, proposalID, page, limit, func(vote govtypes.Vote) bool {
 		votes = append(votes, vote)
 		return false
 	})
@@ -93,9 +118,9 @@ func (k Keeper) GetVotesPaginated(ctx sdk.Context, proposalID uint64, page, limi
 }
 
 // GetVote gets the vote from an address on a specific proposal.
-func (k Keeper) GetVote(ctx sdk.Context, proposalID uint64, voterAddr sdk.AccAddress) (vote types.Vote, found bool) {
+func (k Keeper) GetVote(ctx sdk.Context, proposalID uint64, voterAddr sdk.AccAddress) (vote govtypes.Vote, found bool) {
 	store := ctx.KVStore(k.storeKey)
-	bz := store.Get(govTypes.VoteKey(proposalID, voterAddr))
+	bz := store.Get(govtypes.VoteKey(proposalID, voterAddr))
 	if bz == nil {
 		return vote, false
 	}
@@ -105,43 +130,43 @@ func (k Keeper) GetVote(ctx sdk.Context, proposalID uint64, voterAddr sdk.AccAdd
 }
 
 // setVote set a vote.
-func (k Keeper) setVote(ctx sdk.Context, vote types.Vote) {
+func (k Keeper) setVote(ctx sdk.Context, vote govtypes.Vote) {
 	store := ctx.KVStore(k.storeKey)
 	bz := k.cdc.MustMarshal(&vote)
 	addr, err := sdk.AccAddressFromBech32(vote.Voter)
 	if err != nil {
 		panic(err)
 	}
-	store.Set(govTypes.VoteKey(vote.ProposalId, addr), bz)
+	store.Set(govtypes.VoteKey(vote.ProposalId, addr), bz)
 }
 
 // SetVote set a vote.
-func (k Keeper) SetVote(ctx sdk.Context, vote types.Vote) {
+func (k Keeper) SetVote(ctx sdk.Context, vote govtypes.Vote) {
 	k.setVote(ctx, vote)
 }
 
 // GetVotesIterator returns an iterator to go over all votes on a given proposal.
 func (k Keeper) GetVotesIterator(ctx sdk.Context, proposalID uint64) sdk.Iterator {
 	store := ctx.KVStore(k.storeKey)
-	return sdk.KVStorePrefixIterator(store, govTypes.VotesKey(proposalID))
+	return sdk.KVStorePrefixIterator(store, govtypes.VotesKey(proposalID))
 }
 
 // GetVotesIteratorPaginated returns an iterator to go over
 // votes on a given proposal based on pagination parameters.
 func (k Keeper) GetVotesIteratorPaginated(ctx sdk.Context, proposalID uint64, page, limit uint) sdk.Iterator {
 	store := ctx.KVStore(k.storeKey)
-	return sdk.KVStorePrefixIteratorPaginated(store, govTypes.VotesKey(proposalID), page, limit)
+	return sdk.KVStorePrefixIteratorPaginated(store, govtypes.VotesKey(proposalID), page, limit)
 }
 
 // deleteVote delete a vote for a proposal.
 func (k Keeper) deleteVote(ctx sdk.Context, proposalID uint64, voterAddr sdk.AccAddress) {
 	store := ctx.KVStore(k.storeKey)
-	store.Delete(govTypes.VoteKey(proposalID, voterAddr))
+	store.Delete(govtypes.VoteKey(proposalID, voterAddr))
 }
 
 // DeleteAllVotes deletes all votes for a proposal.
 func (k Keeper) DeleteAllVotes(ctx sdk.Context, proposalID uint64) {
-	k.IterateVotes(ctx, proposalID, func(vote types.Vote) bool {
+	k.IterateVotes(ctx, proposalID, func(vote govtypes.Vote) bool {
 		addr, err := sdk.AccAddressFromBech32(vote.Voter)
 		if err != nil {
 			panic(err)
@@ -152,13 +177,17 @@ func (k Keeper) DeleteAllVotes(ctx sdk.Context, proposalID uint64) {
 }
 
 // IterateAllVotes iterates over the all the stored votes and performs a callback function.
-func (k Keeper) IterateAllVotes(ctx sdk.Context, cb func(vote types.Vote) (stop bool)) {
+func (k Keeper) IterateAllVotes(ctx sdk.Context, cb func(vote govtypes.Vote) (stop bool)) {
 	store := ctx.KVStore(k.storeKey)
-	iterator := sdk.KVStorePrefixIterator(store, govTypes.VotesKeyPrefix)
+	iterator := sdk.KVStorePrefixIterator(store, govtypes.VotesKeyPrefix)
 
 	defer iterator.Close()
 	for ; iterator.Valid(); iterator.Next() {
+<<<<<<< HEAD
 		var vote types.Vote
+=======
+		var vote govtypes.Vote
+>>>>>>> 6f4b45bce5f277e193c4116dbea18212f40e242a
 		k.cdc.MustUnmarshal(iterator.Value(), &vote)
 
 		if cb(vote) {
@@ -168,12 +197,16 @@ func (k Keeper) IterateAllVotes(ctx sdk.Context, cb func(vote types.Vote) (stop 
 }
 
 // IterateVotes iterates over the all votes on a given proposal and performs a callback function.
-func (k Keeper) IterateVotes(ctx sdk.Context, proposalID uint64, cb func(vote types.Vote) (stop bool)) {
+func (k Keeper) IterateVotes(ctx sdk.Context, proposalID uint64, cb func(vote govtypes.Vote) (stop bool)) {
 	iterator := k.GetVotesIterator(ctx, proposalID)
 
 	defer iterator.Close()
 	for ; iterator.Valid(); iterator.Next() {
+<<<<<<< HEAD
 		var vote types.Vote
+=======
+		var vote govtypes.Vote
+>>>>>>> 6f4b45bce5f277e193c4116dbea18212f40e242a
 		k.cdc.MustUnmarshal(iterator.Value(), &vote)
 
 		if cb(vote) {
@@ -184,12 +217,16 @@ func (k Keeper) IterateVotes(ctx sdk.Context, proposalID uint64, cb func(vote ty
 
 // IterateVotesPaginated iterates over votes on a given proposal
 // based on pagination parameters and performs a callback function.
-func (k Keeper) IterateVotesPaginated(ctx sdk.Context, proposalID uint64, page, limit uint, cb func(vote types.Vote) (stop bool)) {
+func (k Keeper) IterateVotesPaginated(ctx sdk.Context, proposalID uint64, page, limit uint, cb func(vote govtypes.Vote) (stop bool)) {
 	iterator := k.GetVotesIteratorPaginated(ctx, proposalID, page, limit)
 
 	defer iterator.Close()
 	for ; iterator.Valid(); iterator.Next() {
+<<<<<<< HEAD
 		var vote types.Vote
+=======
+		var vote govtypes.Vote
+>>>>>>> 6f4b45bce5f277e193c4116dbea18212f40e242a
 		k.cdc.MustUnmarshal(iterator.Value(), &vote)
 
 		if cb(vote) {