+++ x/oracle/keeper/keeper.go
@@ -14,11 +14,16 @@ type Keeper struct {
 	bankKeeper    types.BankKeeper
 	distrKeeper   types.DistrKeeper
 	stakingKeeper types.StakingKeeper
+	CertKeeper    types.CertKeeper
 	paramSpace    types.ParamSubspace
 }
 
 func NewKeeper(cdc codec.BinaryCodec, storeKey sdk.StoreKey, authKeeper types.AccountKeeper, distriKeeper types.DistrKeeper,
+<<<<<<< HEAD
 	stakingKeeper types.StakingKeeper, bankKeeper types.BankKeeper, paramSpace types.ParamSubspace) Keeper {
+=======
+	stakingKeeper types.StakingKeeper, bankKeeper types.BankKeeper, certKeeper types.CertKeeper, paramSpace types.ParamSubspace) Keeper {
+>>>>>>> 6f4b45bce5f277e193c4116dbea18212f40e242a
 	return Keeper{
 		cdc:           cdc,
 		paramSpace:    paramSpace,
@@ -27,6 +32,7 @@ func NewKeeper(cdc codec.BinaryCodec, storeKey sdk.StoreKey, authKeeper types.Ac
 		distrKeeper:   distriKeeper,
 		stakingKeeper: stakingKeeper,
 		bankKeeper:    bankKeeper,
+		CertKeeper:    certKeeper,
 	}
 }